feat(ssd_detection): requant 定点参数软件预转（scale 区每通道 4 字，FPGA 无浮点）
- conv_op.cc：初始化 scale 时按 ref_int8.quantize_params 语义预转
  mult/bias_int/shift/rcl6（round-half-away，double 运算，500 轮随机与 Python 参考一致）
- intelfpga.cc：scale 区 memcpy 由 2×out_c float 改为 4×out_c int32
- intelfpga.h：scale 指针类型 float* → int32_t*，注释说明 4 字布局

+++ de10_nano/ssd_detection/include/intelfpga.h
@@ -171,7 +171,7 @@ struct FpgaConvParam: public NodeParam {
 	int8_t*            d_ka{nullptr}; // device kernel address, [Co/INPUT_CHANNLE_TILE,Ci,Hk,Wk,INPUT_CHANNLE_TILE]
 	int8_t*            oa{nullptr}; // output address, [N,Co,Ho,Wo]
 	int8_t*            d_oa{nullptr}; // output address, [N,Co/INPUT_CHANNLE_TILE,Ho,Wo/INPUT_CHANNLE_TILE]
-	float *			   scale{nullptr}; // filter scale
+	int32_t *		   scale{nullptr}; // 定点 requant 参数，每通道 4 字：mult/bias_int/shift/rcl6（软件预转）
 	struct parameter param;
 };
